+++ CHANGELOG.md
@@ -5,11 +5,11 @@ Note that since we don't clearly distinguish between a public and private interf
 
 ## [Unreleased]
 - Remove `xhr2` dependency for NodeJS, use `fetch`
+- Add `mvs-stories` app included in the `molstar` NPM package
+  - Use the app in the corresponding example
 
 ## [v4.16.0] - 2025-05-20
 - Load potentially big text files as `StringLike` to bypass string size limit
-- Add `mvs-stories` app included in the `molstar` NPM package
-  - Use the app in the corresponding example
 - MolViewSpec extension:
   - Load single-state MVS as if it were multi-state with one state
   - Merged `loadMVS` options `keepCamera` and `keepSnapshotCamera` -> `keepCamera`